requirements.txt
pandas~=2.2.3
numpy~=2.0.2
scikit-learn~=1.6.1
flask
fastapi~=0.115.12
uvicorn
PyYAML
pydantic~=2.11.3
lightgbm